refactored EmployeeInfoCard.vue and formatted emmployee page
